Halt worker threads in a more deterministic fashion
Previously, we just had the sync thread tell all of the threads to
stop at once and then just wait for them all to exit.

This was starting to introduce deadlock scenarios on exit where
a producing thread would be blocked trying to write its final
messages to a full queue when the corresponding consumer had
already exited.

Now we halt threads in batches based on their role, with
upstream producers being halted first followed by their consumers.
This prevents us from halting a consumer before the producer has
stopped and should greatly reduce the frequency of collectors
hanging on exit.

+++ src/collector/collector_sync.h
@@ -47,11 +47,11 @@ typedef struct colsync_data {
     collector_identity_t *info;
     pthread_rwlock_t *info_mutex;
 
-    int pubsockcount;
-    int forwardcount;
-    int emailcount;
-    int gtpcount;
-    int sipcount;
+    size_t pubsockcount;
+    size_t forwardcount;
+    size_t emailcount;
+    size_t gtpcount;
+    size_t sipcount;
 
     void **zmq_pubsocks;
     void **zmq_fwdctrlsocks;
